refactor: move all cli commands in a commands package
