+++ cmd/cli/kubectl-kyverno/commands/apply/.gitignore
