+++ cmd/cli/kubectl-kyverno/commands/apply/log.go
