+++ cmd/cli/kubectl-kyverno/commands/create/templates/metrics-config.yaml
