+++ cmd/cli/kubectl-kyverno/commands/create/templates/test.yaml
