+++ cmd/cli/kubectl-kyverno/commands/create/values/command.go
@@ -5,8 +5,8 @@ import (
 	"strings"
 	"text/template"
 
-	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/create/templates"
-	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/test/api"
+	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/commands/create/templates"
+	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/commands/test/api"
 	"github.com/spf13/cobra"
 )
