+++ cmd/cli/kubectl-kyverno/commands/fix/test/load.go
@@ -4,7 +4,7 @@ import (
 	"os"
 	"path/filepath"
 
-	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/test/api"
+	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/commands/test/api"
 	"k8s.io/apimachinery/pkg/util/yaml"
 )
