+++ cmd/cli/kubectl-kyverno/commands/jp/query/command.go
