+++ cmd/cli/kubectl-kyverno/main.go
@@ -5,28 +5,15 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/apply"
-	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/create"
-	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/docs"
-	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/fix"
-	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/jp"
-	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/oci"
-	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/test"
-	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/utils/experimental"
-	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/version"
+	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/commands"
 	"github.com/kyverno/kyverno/pkg/logging"
 	"github.com/spf13/cobra"
 )
 
 func main() {
-	cli := &cobra.Command{
-		Use:   "kyverno",
-		Long:  "To enable experimental commands, KYVERNO_EXPERIMENTAL should be configured with true or 1.",
-		Short: "Kubernetes Native Policy Management",
-	}
-	configureLogs(cli)
-	registerCommands(cli)
-	if err := cli.Execute(); err != nil {
+	cmd := commands.Command()
+	configureLogs(cmd)
+	if err := cmd.Execute(); err != nil {
 		os.Exit(1)
 	}
 }
@@ -38,21 +25,4 @@ func configureLogs(cli *cobra.Command) {
 		os.Exit(1)
 	}
 	cli.PersistentFlags().AddGoFlagSet(flag.CommandLine)
-}
-
-func registerCommands(cli *cobra.Command) {
-	cli.AddCommand(
-		apply.Command(),
-		create.Command(),
-		docs.Command(cli),
-		jp.Command(),
-		test.Command(),
-		version.Command(),
-	)
-	if experimental.IsExperimentalEnabled() {
-		cli.AddCommand(
-			fix.Command(),
-			oci.Command(),
-		)
-	}
 }
